requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
django-extensions==3.2.3
gunicorn==21.2.0
idna==3.10
kombu==5.5.4
packaging==25.0
psycopg2-binary==2.9.9
prompt_toolkit==3.0.52
Pillow==10.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==5.0.4
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.4
